efi: add --disable-winpv option
Add a --disable-winpv option to bareflank.efi, that when given
will disable the code needed to support the Windows PV drivers.
This option implies --disable-xen-pfd, so the xenbus driver is
not loaded by windows.

This change enables Windows PV operation by default when running
bareflank.efi, which is why --disable-winpv is added in the first
place. This means the default bareflank.efi will assume the
Windows PV drivers are installed and that Windows is the boot target.
If you want to boot into Linux instead, pass --disable-winpv on
the EFI command line (and set EFI_BOOT_NEXT in your config.cmake
as usual).