deploy/windows: start using components in app.iss
Start using 'Components:' directives to enable the end-user to select
what they want the installer to place on the system. This defines two
installation flavors: standard and custom.

The standard install is what is required for a complete, working system
and is selected by default. The custom install can be used to select
what components should be installed, and allows for partial upgrades.
It also enables developers to chose things helpful for debugging and
testing, such as booting into an EFI shell.