vmm: fix context-switching bug in base hypervisor
This fixes a major bug in the context switching assembly glue in the base
hypervisor. The VMM itself uses SSE state and instructions, but was only
saving/restoring XMM0-XMM7 on context switches. The patch includes
XMM8-XMM15 and the MXCSR register. It is likely this hasn't been an
issue in the absence of guest VMs because the VMM (presumably) doesn't
modify XMM8-XMM15 or MXCSR. But the guest VMs (again, presumably) do,
and not saving the upper XMM's can cause state corruption.

Now each entry to the VMM (i.e. vmexit, loader, or exception) saves all
the SSE state. Each entry to a VM (vmlaunch, vmresume, and promotion)
restores it.

There was also bug in start_x64.asm that was not restoring the xmm0-xmm7
registers after returning from _start_c, but rather just moving them back
onto the stack; this patch fixes that as well.