xen/sched: add debug statements for SCHEDOP_shutdown
Add some print statements in SCHEDOP_shutdown handler that prints
the shutdown reason (not all are crashes). Also return -EINVAL
to the guest and true to the vmcall handler in the VMM. This will
allow us to vmenter back into the guest, which will then (assuming
Linux) BUG(), giving us a nice stack trace to work with.