deps/hypervisor: fix register offset bug in exception dump
Update the register offsets in the view to ensure that rip and rsp
are printed correctly. Also, since both error-code and non-error-code
exception handlers have the same stack layout now, make sure the offsets
into the view are the same.

As a sanity check, add an ensures() in the vcpu.cpp that checks that the
pointer written into the IST entry of the TSS is 16-byte aligned. This
is required because the movdqa instructions used on the xmm registers
will fault if the address isn't aligned.