Addresses issues with state
This patch fixes crashing issues with the hypervisor. Specifically,
CR2 was being leaked between guest VMs, which would cause a segfault
if both the root OS and the guest OS were in the middle of servicing
a page fault. As a result, we not only cleaned up issues with CR2
but with all of the other state registers on the system.