deps/hypervisor: save/restore cr2 on vmexit/vmentry
Add CR2 into the vcpu_state_t and save and restore its value on each vmx
transition. This fixes an issue in guest VMs that was causing segfaults.

Whenever a page fault occurs, hardware updates CR2 with the faulting
address. If an interrupt occurs while a guest vcpu is running, an
external-interrupt exit traps to the vmm, which then injects the
interrupt into the root domain and resumes back into uvctl so that the
interrupt can be handled by the root domain's OS.

If an interrupt occurs after a page fault in the guest VM but before
Linux has read CR2 *and* CR2 changes after the vmm resumes into uvctl
(because e.g. another process is scheduled that accesses an address
which faults), then once the guest is resumed into again, its page
fault handler will read the CR2 of the fault that occured in the root
domain, which will cause it to walk an incorrect path along its page tables.

Simply saving and restoring CR2 on each exit/entry fixes the problem.