bfexec: Add locking of UART buffer
For the dump uart hypercall, the vmm maps in the guest virtual address
of the buffer passed in from bfexec before writing data to it. In order
for the map to resolve properly, the address needs to be populated in
bfexec's page tables.

Ensure this is the case by using VirtualAllocEx/VirtualLock on Windows
and mmap/mlock on Linux. Also provide wrappers 'alloc_locked_buffer' and
'free_locked_buffer' around the platform-specific calls.