deps/hypervisor: Fix m_eptp and relax single-context requirement
This patch ensures that the m_eptp member of the ept_handler class is
passed to every function in the ept_pointer vmcs field intrinsic (it was
missing for the phys_addr). It also relaxes the requirement on
single-context invept. If it is not available, the global invept is used
instead of throwing an exception.