vmm: Isolate IA32_TSC_AUX msr
Since RDTSCP is passed-through to guests, the IA32_TSC_AUX msr needs to
be isolated rather than emulated so that RDTSCP returns the correct
value of the IA32_TSC_AUX msr. This msr is special because it can be
read via rdmsr, rdtscp, and rdpid. Each of these instructions return the
current value in hardware, so either you have to emulate all of them or
pass-through all of them. Since we pass them through, the value of
IA32_TSC_AUX in hardware must be correct at all times.