vmm: Add mce handler and move invalid opcode handler
Add a machine-check exception handler in xen_vcpu. It needs to be
in xen_vcpu rather than the microv vcpu so that the xen runstate
can be updated properly before the world switch via return_interrupted().
With the mce handler in place, hide the mce and mca bits from cpuid from
the guest vcpus as well.

Move the invalid opcode handler out of the xen_vcpu and into the microv
vcpu, since its operation is not dependent on the xen_vcpu in any way.