vmm: Add nmi handler for guest xen_vcpus
Before this patch, if an NMI and/or NMI exit occured in guest vcpu
context, it would have been injected into the guest vcpu. This is
problematic because we dont expose any NMI sources to the guest kernel.

Add an NMI handler for xen vcpus that simply queue_nmi() then world
switch into its root vcpu so that the root OS can handle the NMI as needed.