deploy/windows: Update app.iss
This patch removes the explicit listing of each component to declutter
the install options that are presented to the end user. It also removes
the EFI shell as an install option to keep it simple. If testers or
developers needs to pass command line arguments to the VMM, they will
need to add the shell themselves.

The other major change moves the code for checking compatibility to
InitializeSetup() instead of running it at the end of the [Run] section.
With this change, the first thing Setup does is install the VC runtime
libraries needed by the userspace applications. It then allows
powershell scripts to run locally, then executes each compat checker.
The rest of the install proceeds only if all the compat checkers return
successfully.