windows/scripts: Ask for user/pass each startup
Remove the saving of the username and password into a file. Instead
prompt the user for their creds each startup. This is more cumbersome
for the end-user, but avoids an incorrect password from being saved into
a file, and also enables the user to change their password without it
affecting startup.