vifconnect.ps1: Disable DHCPv6 and IPv6 on non-mgmt interfaces
It is possible for the Windows' DHCPv6 client to get stuck in a
loop of Solicit messages. When this happens, the messages will be sent
to the backend, which currently drops all IPv6 packets. As such the
Windows client never hears a response, which halts traffic related
to IPv4 as well.

This patch disables DHCPv6 and IPv6 completely on non-management
interfaces to avoid this bug.