drivers/winpv: Fix xenstore shutdown/restart hangs
Add a mechanism for detecting that the xenstore backends are dying.
When the backends are dying (i.e. on shutdown), any xenstore request
will hang because no response will be received. This patch adds
XENBUS_STORE_{SET,GET}_BACKENDS_DYING so that clients of
XENBUS_STORE can set/get the "dying" state of the backends.

This is used by the xenbus FDO whenever the ioctl opened by uvctl is closed.
Any subsequent store access will see that the backends are dying and
the loop in StoreSubmitRequest will terminate.