xenvif: More fixes for more hangs
This patch checks for BackendsDying in the transmitter ring watchdog
thread, and if TRUE, disables the ring. This fixes a hang on a
particular Dell XPS that happens every time the system is
shutdown/restarted.

To support the subsequent paths that are taken after the ring is
disabled, the FrontendSetState and FrontendWaitForBackendXenbusStateChange
functions check BackendsDying and bail if TRUE after setting the state
to FRONTEND_UNKNOWN.

Also, NDIS sends a power down and power up IRP on the restart path.
Because the frontend's state is fixed to FRONTEND_UNKNOWN once the
backends are detect as dead, the D3 to D0 transition will fail, which
triggers an ASSERT in __PdoSetDevicePower. This ASSERT is modified here
to only trigger if BackendsDying is FALSE.