app.iss: Check for install before compat checks
Check if the Beam_is1 registry key exists to determine if a prior
install succeeded. If the key exists, then the compat checks must have
run prior, so the compat checks in that case. This is needed to support
upgrades while Beam is running.

Also install the PNP drivers with /f to force install irrespective of
version numbers, and reorder the winpv drivers to install xennet then
xenvif then xenbus. This allows xennet and xenvif to be upgraded before
xenbus in upgraded, otherwise the system will be prompted to reboot before
xennet and xenvif are ready.