vmm: Add TLB shootdown support
This patch adds the ability to perform a TLB shootdown for invalidating
stale EPT entries. To support this, an IPI mechanism is added to send
events to the VMM running on a different core(s). It uses the local APIC
to send an all-excluding-self INIT interrupt. When the INIT signal
arrives at a remote processor, it traps to the VMM (or is held pending
and then traps on vmentry if the VMM happens to be running when the
signal arrives).

From there the INIT signal exit handler checks an IPI code. If the code
is zero, it is assumed an actual INIT was sent from the root OS as apart
of AP bootup. If the code is non-zero, it is passed to handle_ipi which
dispatches it appropriately. IPI_CODE_TLB_SHOOTDOWN is the only code
defined so far.