vmm: Add struct page_range
This structure is used to represent a contiguous range of pages. It is
useful for effecient bookeeping of donated pages when large regions of
the donation are contiguous. Instead of keeping a list of every page's
address, the range is maintained. From initial testing, the range-based
approach saves an order of magnitude of memory on Windows and three
orders on Linux.