vmm: Isolate guest memory from root domain via donation
This patch fully implements "donation" of memory from the root domain to
a guest domain. Donation in this sense means that the memory is unmapped
from the root domain's EPT and is mapped into the guest domain's EPT
with caller-specified permissions. Whenever the guest is destroyed, the
root domain is allowed to reclaim the memory.