builder: Add retry logic to page donation
The __domain_op__donate_page* hypercalls return an AGAIN error code to
indicate that the operation could not be completed at this time and to
try again if desired. This change modifies the __domain_op__donate_page*
callsites to sleep for at least 100 microseconds after receiving an
AGAIN code, with up to 3 retries. These limits are arbitrary and can be
adjusted later if need be.