depends: Add capstone disassembler
Add capstone to support disassembly. Sometimes instructions need to be
disassembled to support emulation of guest accesses to various memory
regions (ECAM, xAPIC, etc.). This library provides a well-tested and
reasonably efficient means for supporting this need. It also supports
many other architectures other than x86, so it can be used in the future
for them if needed.