vmm/pci: Stop hiding the MCFG table
This change disables the hiding of the MCFG table from the root vm,
so that it may use PCI-e features of devices that are not being
passed-through to a guest.

Remove the ECAM-related data and functions of struct pci_dev, since the
ECAM of passthrough devices is no longer being remapped to m_vcfg.