vmm: Add IO bitmap shootdown
This patch adds another shootdown reason - updating the IO bitmaps used
by vcpus. Section 24.11.4, Vol. 3 of the Intel manual states that memory
referenced by vmcs fields such as the IO and MSR bitmaps must only be
updated if all vcpus that use them are *not* running in VMX non-root
mode.

To ensure this is the case, an IPI is sent with code
IPI_CODE_SHOOTDOWN_IO_BITMAP that forces each CPU into VMX root mode,
where they all wait until the sender of the IPI is done with the update.
Note that unlike the TLB shootdown, the remote CPUs dont need to do any
invalidation; they just wait for the update to complete and return.