vmm/pci: Unmap passthrough BARs from root vm
This patch adds support for unmapping a passthrough device's BARs from
the root vm. To do this, each passthrough device's BARs are parsed on
vcpu0 via parse_bars() called in probe_bus(). Then when the root config
handlers are installed in add_root_handlers (which is called once per
root vcpu), the BARs are unmapped.

The root_cfg_out handler is modified to detected BAR relocations so that
the mapping permissions can be updated appropriately. It does this by
checking if the IO and/or Memory space enable bits are set on each write
to the device's command register (bytes [5:4] in configuration space).
If they are set and the base addresses have changed, then
relocate_pmio_bars and relocate_mmio_bars are called as appropriate.
Both of these calls cause a shootdown, so they have "retry" semantics.
If the shootdown is occupied (because some other core is trying to
shootdown), then the 'again' field is set to true in the 'info' parameter,
and the call returns to try again. This repeats until the shootdown
succeeds.

Whenever the guest handlers are added with add_guest_handlers(),
the current value of the BARs are mapped into the guest vm, like normal.

To support the mapping/unmapping of BARs, helper functions are for
PMIO and MMIO.

Finally, the root vcpus' EPT violation handler is updated to ensure that
accesses to a device's BARs log a warning and skip the instruction (also
while here, accesses to donated pages are also logged and skipped).