vmm/ept: Add flush_tables
This function can be called by users of the EPT mmap to ensure that all
the paging structures are written back to memory. This is useful when
the EPT is being shared with an IOMMU with an incoherent page walk and
the user of the mmap needs to ensure that the tables are written to
memory.

This function accepts a boolean snoop_ctl that if true, will set the SNP
bit (bit 11) in each leaf entry before writing the entry back to memory.