vmm/pci: Split device assignment and mapping
Add assign_pci_device to the microv domain. This adds the given device
to a per-domain structure, and can be called at various times depending
on the domain. For example, it is called after the VMM is unmapped
for the root domain, and for a xen domain with passthrough (e.g. NDVM)
it is called in the assign_device domctl hypercall.

After all the devices have been assigned to a domain, then
prepare_iommus() is called to resolve the coherency and snoop control
attributes of each IOMMU that references the domain's EPT. These
values are then set in the domain's EPT mmap, and if necessary,
the mmap's tables are flushed and the snoop control bit is set.

After the domain's IOMMU's are prepared, DMA remapping is enabled
for each device via map_dma. This program's each device's context
entry with the value of the domain's EPTP and sets the translation
type to Untranslated.