vmm: Add domain IOTLB invalidation
Make sure that IOTLBs are properly invalidated whenever the EPT
that they reference is modified. This change focues on domain-selective
invalidation (DSI), and specifically performs DSI for the root
domain whenever:

  1) PCI passthrough device BARs are relocated
  2) Guest domains with root-donated pages are preparing to run

This patch does not perform any DSI for guest domains with passthrough
devices, because the current code expects() that CAP.CM is 0
in each IOMMU (meaning that non-present entries are not cached), and
therefore no entries exist for invalidation in the first place.