vmm/pci: Disable passthrough for devices with an RMRR
If a PCI device has an RMRR, then that region specified in the RMRR must
be also be communicated as reserved to the guest that the device is
passed-through to. Since we dont handle this right now, if a device with
an RMRR is found, just disable passthrough of that device. In practice,
this means that USB host controller and GPUs will not be able to be
passed through properly (based on systems that have been tested so far).