xen/vcpu: Evade NVIDIA virt detection
NVIDIA drivers detect is they are running in a virtual environment
be reading leaves 0x40000000 and 0x40000100, and refuse to start
the device if they are.

To evade this, return zeros for this leaves on root vcpus, and modify
the Windows PV driver to assume the base Xen leaf is 0x40000100
without actually reading the leaf.