vmm/iommu: Add map_bus
Add an iommu::map_bus function that associates a context table with a
domain. Each entry in the context table is programmed with Untranslated
translation type with the domain's EPT pointer as the second-level
page-table pointer. Finally the root table entry is programmed with the
physical address of the context table.

Note that since every context entry is marked present, all current and
(future) hotplugged devices will use the domain's EPT for DMA translation.