vmm/xen: Map/unmap xenbus fdo hole pages
The Windows PV xenbus FDO code allocates a 2MB, physically-contiguous
region of memory (called a "hole" in the driver) from Windows on startup,
then unmaps the region from its EPT with XENMEM_decrease_reservation. It
then divvies this region up in 4K chunks for various purposes. The
shared info page is the first 4K. The grant table pages follow after that.
And the gpa's used for GNTTABOP_{map,unmap}_grant_ref are allocated from that
region as well.

To properly support the GNTTABOP hypercalls made from xeniface, the
gpa's need to be mapped to some other page from the granting domain's
EPT, i.e., not all gpa's from the hole are identity mapped.

Modify the hypercalls involved with mapping/unmapping memory in the
FDO hole to make sure they actually do the map/unmap (before they
just kept everything identity mapped, but that no longer is sufficient
because of the GNTTABOPs).