xen/gnttab: Add support for map/unmap grant ref on Windows
Relax some assertions in the map/unmap grant ref paths that allow the
hypercalls to be made from the (Windows) root domain. Add a check that
ensures that the local gpa provided in the map/unmap arg is in the FDO
hole created by the FDO driver.

Also, dont invept or invalidate the IOTLB in the map path, because 1) the
permissions are moving from not-present to present and 2) the current
IOMMU implementation ensures that the IOMMU hardware doesn't use CM (so
the hardware doesn't cache non-present entries).