winpv/xeniface: Pin cancelled gnttab IRPs
On shutdown and restart, unmap IRPs from xeniface get cancelled
and run on various different cores. This is problematic as
the core on which the original map/access was done may be different
than the unmap, and the wrong TLB is be invalidated in this case.

Add a ProcNumber to the xeniface context that is recorded on
each MapForeignPages. Then in the cancelled IRP handler, set the
thread's affinity with this ProcNumber before doing the unmap
hypercall.