xen: Fix gnttab and evtchn issues found during HLK testing
Prior to this patch, the winpv paths in the evtchn and gnttab code
assumed that the PV drivers are started and stopped only once over
the course of the VMMs execution. The HLK tests routinely
start/stop/remove the drivers many times after the VMM itself has
started.

Ensure that the evtchn and gnttab data structures cope with
hypercalls coming from the Windows PV init routines many times
while the hypervisor is running.