vmm: Add __enum_run_op__notify_domain return code
This code is returned to the root vcpu whenever a guest uses either the
EVTCHNOP_send hypercall to notify a remote domain or when a physical
interrupt arrives bound to a passthrough device assigned to a guest
domain.

In the first case, the xen domain ID is converted into a microv domain
ID. In the second case, the microv domain ID is derived from the
interrupt that arrived. In both cases, the microv domain ID is returned
to the root vcpu that will use the ID to lookup the root-specific event
handle to signal before yielding execution.