xenbus: Add two new ioctls
Add IOCTL_XENBUS_ADD_USER_EVENT and IOCTL_XENBUS_SET_BACKEND_STATE.
The former is used to register a userspace event handle with xenbus.
The handle is signaled by the driver each time EvtchnSend is called.
The latter is for telling xenbus that the service VM that is hosting
dom0 is dying, and that it shouldn't wait forever for xenstore IO
to complete.