xenbus: Implement event and backend state ioctls
Implement IOCTL_XENBUS_ADD_USER_EVENT. This adds the provided
userspace handle to a list that is checked every time a new event
channel is opened. If the channel has a matching remote domain ID,
the event is signaled on every EvtchnSend.  Note this adds a function to
the evtchn_interface.h, so those files are updated in xenvif and xeniface
as well.

Implement IOCTL_XENBUS_SET_BACKEND_STATE. This is used to signal
to xenbus that the service VM hosting dom0 is dying, so it should not
expect xenstore IO to succeed anymore. Also remove the OB callbacks
that were used to protect uvctl. This is not really needed anymore,
since it runs as a system service that only implements PRESHUTDOWN,
so "typical" users shouldn't be able to turn it off.