uvctl: Rework ioctls and events
Dont hold onto the xenbus_fd for the duration of execution. This
isn't necessary and can actually cause HLK tests to fail since they
are not able to remove the xenbus device if it has open handles.
Instead, just open it when the process is terminating and
perform a IOCTL_XENBUS_SET_BACKEND_STATE ioctl to tell xenbus that
the backend is dying.

Whenever a vcpu is launched, open an event handle and give it
to xenbus via IOCTL_XENBUS_ADD_USER_EVENT. If the vcpu is part
of the NDVM, then additionally give the handle to visr via the
IOCTL_VISR_REGISTER_EVENT ioctl. In all cases, whenever a driver
sends an event of interest, it signals the event so that the
WaitForSingleObject call wakes up, prompting recipient vcpu to
be scheduled. This same event handle is also signaled when
return_notify_domain is received from a guest vcpu.