uvctl: Fix OOB read in event signaling mechanism
Prior to this patch, init_events wasn't being called, so the event vector
was not being initialized, therefore when an event signal occured, it was
accessing garbage memory.

Fix this by actually calling init_events. Also start using a static
array with MAX_DOMAINS entries for events. The current value of
MAX_DOMAINS is 16, which is 4 times as many we use now.

Add two helpers for init_events, init_xenbus_events and
init_visr_events. Each of these opens the respective driver's IOCTL
interface and performs the necessary ioctl to add the domain's event
handle to the driver.