xenbus: Remove xenfilt
xenfilt is used for PnP of emulated devices, which we dont use.
It also is responsible for the uninstall bug that causes Windows
to fail booting, saying that boot critical files are missing.

This patch removes xenfilt and all references to it throughout
the rest of xenbus and the build system.