CI: Fix issue when building for the first time
On the first run, the build and cache directory won't exist in the
cached gitlab directory for this build. The `rm` command will have
the exit code `false` which is picked up by gitlab as a failed command.
The command parenthesis will force the return value to true and allow
gitlab to continue.