xen/vcpu: Add missing runstate updates
Add calls to update_runstate when the guest is preempted
back into the root vcpu. The runstate entry times are tracked
by the linux scheduler to properly account for stolen time.

Also while we're here, remove the calls to wmb in update_runstate.
These are really only needed if it is possible for the linux guest
to run at the same time as the update, which under the current
implementation is not possible. Note the root domain doesn't
use the runstate information at all, so this only applies to
the linux guests.