cmake: Add .gitsubtrees for clang-format
The previous method used `git log` to get the list of git subtree paths
which requires cloning the entire git history to work.

Instead, we now use a `.gitsubtrees` file in which each line is a
relative path to a subtree. These paths are used to exclude them from
`git ls-files` which in turn is used to the actual list of files to
format.