xen/evtchn: Add register-based EVTCHNOP_send
Add a cmake config option XEN_REGISTER_SEND that when enabled tells the
evtchn code to assume the port of the EVTCHNOP_send hypercall is passed
directly in a register rather than in a evtchn_send_t.

This avoids the need to map in the evtchn_send_t into the VMM, and
thus saves 5 EPT walks and 5 invlpgs per send. This shaves off roughly
half the cost of the send hypercall.

The default value of XEN_REGISTER_SEND is OFF. When it is ON, and
you want to use the Windows PV drivers, you must also pass the
'-registersend' option to xenbus/build.ps1.