vmm: Fix TLB invalidation bug
The VMM has always used a single set of page tables that is shared with
every CPU. Prior to this change, whenever a virtual address mapping was
freed, invlpg page was used to flush the TLB.  However the other cores
were not shot down, and so the invlpg was only done locally. As soon as
the virtual mappings A/D bits are set, other cores' TLB prefetch
mechanisms are allowed to pull in the entry. Then when that same virtual
address is subsequently mapped on some remote core, the core uses the
stale physical address from the original mapping.

Fix this issue by moving the invlpg to the allocation of the mapping
rather than the freeing.