builder: Misc fixups for WHQL signing
Add builder.rc that specifies driver properties like copyright info,
Product Name/Version, etc. Ensure it is compiled by adding
ResourceCompile element in builder.vcxproj.

Generate version.h (which is used by builder.rc) from version.tmpl
from the genfiles.ps1 script.

Disable test signing in builder.vcxproj to ensure that the WDK test
certificate is not embedded in builder.sys. Note that package.vcxproj
*does* have test signing enabled, so the cat file (and only the cat
file) will be signed by the test cert. This enables the driver to
still be used for testing purposes (the entire package is considered
signed if the cat file is signed) and for production (Microsoft includes
a signed cat file as part of the WHQL signing process, so the test
cat need not be included).