winpv: Disable test signing in driver .vcxprojs
Each driver component with a vcxproj should have SignMode set
to Off. This prevents the test cert from signing the driver's binary
files. Note that each package.vcxproj still has test signing enabled
so that the test-signed cat file can be used for testing the drivers
Note though that this file would not be submitted to the WHQL.