Update MicroV Hypercall Specification.md
Updates the specification to add extra support for
mv_pp_op_{msr,cpuid}_get_{supported,emulated,permissable}_list.

While we are still supporting the original behavior, we can now also
optionally remove the need for the caller to provide and maintain a list
of registers normally needed to find out which ones are supported in the
list. Instead when MV_{RDL,CDL}_FLAG_ALL is given, the entire list is
returned via the shared page.

This modification also includes a mechanism to handle a list that would
be larger than what a single shared page can hold.