VMM: Add mv_pp_op_cpuid_get_supported_list support
- add support for mm_pp_op_cpuid_get_supported_list
- remove support for MV_CDL_FLAG_ALL
- add integration test

The MV_{R,C}DL_FLAG_ALL flags were originally added for
mv_pp_op_msr_get_supported_list in order to prevent the caller from
having to hardcode a list which might be out of sync with the
hypervisor. For CPUIDs, we don't have this issue since the list of
supported CPUIDs can be retrieved via Fn0000_0000[0][EAX] and
Fn8000_0000[0][EAX].

The integration test for mm_pp_op_cpuid_get_supported_list checks the
entire list of supported CPUIDs against a known list.
For troubleshooting or for the purpose of a demo, the following flags
can be (un)commented in the integration test:

```
flags |= integration::CPUID_PRINTER_FLAG_PRINT_ERROR;
// flags |= integration::CPUID_PRINTER_FLAG_PRINT_SUPPORTED;
// flags |= integration::CPUID_PRINTER_FLAG_PRINT_UNSUPPORTED;
```