utils: Add qemu_strace.sh script to trace IOCTLs
This script allows us to trace the IOCTLs made by Qemu on Linux. This is
very useful in order to verify or debug our progress.

The following arguments were added to this script:
- setup (sets up the directory used by the Qemu)
- run (runs Qemu with a simple UEFI VM)

The following environment variables can be overridden:
- STRACE_PATH
- QEMU_PATH
- BUILD_DIR

While running Qemu, a ./qemu/strace.log file will be generated
containing all of the IOCTLs made by Qemu.