shim/linux: add _IOWR_LIST macro
Add a macro, _IOWR_LIST, as a hack to assist with kvm structs that
contain zero-length arrays. We cannot use zero-length arrays in this
project without many consequences in testing, clang-tidy, etc. However,
in these cases, we can get similiar behavior by using a pointer to the
array element type. For example,

    struct kvm_msr_list {
	__u32 nmsrs; /* number of msrs in entries */
	__u32 indices[0];
    };

translates to

    struct kvm_msr_list {
	uint32_t nmsrs; /* number of msrs in entries */
	uint32_t *indices;
    };

However, this changes the ioctl number calculated with _IOC, and causes
the shim to error out with "invalid ioctl cmd." This is because the
sizeof the two structs are not the same, because a zero-length array has
size zero, while a uin32_t * has size 8 on a 64-bit system.

Re-define KVM_GET_MSR_INDEX_LIST using this macro so that the ioctl
number matches the real KVM implementation.

Signed-off-by: Nick Rosbrook <rosbrookn@ainfosec.com>