shim/linux: fix E2BIG behavior in KVM_GET_MSR_INDEX_LIST
The KVM API states that when the number of MSRs is greater than the
value set in nmsrs by the caller, the kernel will set the correct value
in nmsrs and return -E2BIG. This behavior was excluded from the first
implementation of KVM_GET_MSR_INDEX_LIST. Define SHIM_2BIG so that other
shim functions can implement this behavior, which is a common pattern in
ioctls.

Also, fix the final copy_to_user in dispatch_system_kvm_get_msr_index_list.
Although we use a uint32_t * in our struct definition, the real KVM API
uses a zero-length array. This means that the memory region for indices
is guaranteed to be contiguous with nmsrs, whereas our implementation
requires two separate copies: one to the struct base address, and one to
the address pointed to by indices.

This change means that our integration test is now broken. However, the
QEMU integration test works as expected.

Signed-off-by: Nick Rosbrook <rosbrookn@ainfosec.com>