shim/linux: use normal array in kvm_msr_list
Instead of using a pointer for the indices entry in kvm_msr_list, use a
normal array and define the KVM_MSR_GET_INDEX_LIST ioctl using the
_IOWR_LIST macro (see "shim/linux: add _IOWR_LIST macro").

Define MSR_LIST_MAX_INDICES to be 128, as this leaves some room to keep
the stack frame < 1024 bytes. This may be too low of a max, but is
plenty for our current QEMU and integration testing.

Signed-off-by: Nick Rosbrook <rosbrookn@ainfosec.com>