shim: Fix issues with kvm_get_msrs
We need to use copy_from_user twice, once to get the nmsrs value in
order to calculate the size and a second time with the correct size
to copy the entire list of msrs that the user is requesting.

This patch also fixes copying back to user as well.

It also fixes an error seen from the hypervisor when handling the
mv_vs_op_msr_get_list hypercall that this IOCTL makes. The number of
entries is now correct.

Signed-off-by: Christopher Pelloux <git@chp.io>