shim: Fix kvm_get_msr_index_list continuation case
This patch improves performance and fixes the continuation case for
mv_pp_op_msr_get_supported_list.

The total number of MSRs can be calculated after the first call to
mv_pp_op_msr_get_supported_list, allowing to return right away if the
provided buffer is not large enough to fit all the MSRs which improves
performance.

For the continuation case, mv_rdl_t.reg1 should be set to zero during
initialization. Subsequent calls need to update mv_rdl_t.reg1 with the
index to resume from (the input case).

Whenever mv_rdl_t is used by the shim, and reg0 / reg1 are not used,
reg0 and reg1 should always be set to zero.

The description of mv_pp_op_msr_get_supported_list in the MicroV
specification was reworded slightly in order to clarify the use of
mv_rdl_t.reg1.

Signed-off-by: Christopher Pelloux <git@chp.io>