Fix boot loop / A20 line / emulation of port IO IN
The guest VM is stuck in a boot loop. The last instruction to execute
before a guest reset is a write to port IO 92h which is known as the
A20 Line.

The guest is trying to enable the A20 Line as seen here:

```
// enable a20
inb $PORT_A20, %al
orb $A20_ENABLE_BIT, %al
outb %al, $PORT_A20
```

The actual problem occurs during the emulation of the port IO read
instruction `inb $PORT_A20, %al` in which QEMU tries to set the AL
register to the value 0, but MicroV wrongly sets it to 0x8F. On the
next port IO write instruction, `outb %al, $PORT_A20`, QEMU would then
see the first bit of the A20 Line enabled which is the "reset bit" and
would cause QEMU to reboot the VM.

In order to fix this issue, we need to propagate the emulation done by
QEMU back into MicroV on a port io read intercept (i.e. in{b,w,l}
instructions) and set the guest's RAX register to the correct value, as
is done by this patch.

Signed-off-by: Christopher Pelloux <git@chp.io>