vmm: Higher precision TSC and fix early boot issue
On systems that support it, the TSC will now be calibrated with a
higher precision. This method of calibrating the TSC follows more
closely the Intel manual and should be more robust on newer systems.

The previous method of calibrating the TSC could fail on newer systems,
e.g. on 11th Gen CPU i5-1145G7, which would then prevent the system from
booting. This patch fixes this issue.