cmake: Add ENABLE_SERIAL config
Add the ENABLE_SERIAL config to control the enablement of debug output
over serial. When omitted, the default is ON, same as previously.
However, the CI will now enable serial in the debug build only.

The reason for this switch, is due to some Windows systems having
slowdowns during boot. This temporarily fixes this issue, until more
investigation can be done. Another interesting workaround, that doesn't
require disabling serial in the VMM, is to enable EMS in Windows.